コンパイル通るように各クラス修正、RotationMatrixクラス実装

+++ N2_3DGameLibrary/MathUtil.cpp
@@ -4,7 +4,7 @@
 
 float wrapPi(float theta) {
 	theta += kPi;
-	theta -= floor(theta * k1Over2Pi) * k2Pi;
+	theta -= floorf(theta * k1Over2Pi) * k2Pi;
 	theta -= kPi;
 	return theta;
 }
@@ -16,5 +16,5 @@ float safeAcos(float x) {
 	if (x >= 1.0f) {
 		return 0.0f;
 	}
-	return acos(x);
+	return acosf(x);
 }
